+++ script/deploy.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 ########################################################
-# deploy Folio mod-cataloging
+# deploy Folio mod-marccat
 #
 # Author : Christian Chiama - christian.chiama@atcult.it
 # Version: 1.0