+++ src/main/java/org/folio/cataloging/exception/InvalidShelfListTypeException.java
@@ -1,5 +1,0 @@
-package org.folio.cataloging.exception;
-
-public class InvalidShelfListTypeException extends ValidationException {
-
-}