+++ src/main/java/org/folio/marccat/business/common/DataDigAdminException.java
@@ -1,0 +1,21 @@
+package org.folio.marccat.business.common;
+
+public class DataDigAdminException extends Exception {
+  private static final long serialVersionUID = -8207968816842862351L;
+
+  public DataDigAdminException() {
+    super();
+  }
+
+  public DataDigAdminException(String message, Throwable cause) {
+    super(message, cause);
+  }
+
+  public DataDigAdminException(String message) {
+    super(message);
+  }
+
+  public DataDigAdminException(Throwable cause) {
+    super(cause);
+  }
+}