+++ src/main/java/org/folio/marccat/business/common/OrderConfigurationException.java
@@ -1,0 +1,12 @@
+package org.folio.marccat.business.common;
+
+import org.folio.marccat.exception.ValidationException;
+
+public class OrderConfigurationException extends ValidationException {
+
+  /**
+   *
+   */
+  private static final long serialVersionUID = 1L;
+
+}