+++ src/main/java/org/folio/marccat/exception/InvalidShelfListTypeException.java
@@ -1,0 +1,5 @@
+package org.folio.marccat.exception;
+
+public class InvalidShelfListTypeException extends ValidationException {
+
+}