+++ src/main/java/org/folio/marccat/util/TagConstant.java
@@ -1,0 +1,14 @@
+package org.folio.marccat.util;
+
+public interface TagConstant {
+  public static final String[] AUTHORS_TAG = {"100", "110", "111"};
+  public static final String[] TITLES_TAG = {"130", "240", "245"};
+  public static final String[] PUBLISHERS_TAG = {"260", "264"};
+  public static final String[] PUBLISHERS_TAG_260 = {"260"};
+  public static final String[] PUBLISHERS_TAG_264 = {"264"};
+  public static final String[] EDITIONS_TAG = {"250"};
+  public static final String[] TYPE_MEDIA = {"337"};
+  public static final String[] TYPE_SUPPORT = {"338"};
+
+  public static final String[] ISBN_ISSN_TAG = {"020", "022"};
+}