+++ NEWS.md
@@ -1,3 +1,7 @@
-## 0.1.0 2018-09-30
- * Initial work
- * Basic CRUD interfaces in place
+## 1.1.0 Unreleased
+ * Template Management
+ * Holding Management
+ 
+## 1.0.0 2018-10-31
+ * Initial release
+ * Search Functionality interface